refactor miccha exception
