--- src/main/java/com/miccha/server/exception/DuplicateEmailException.java
@@ -1,10 +1,0 @@
-package com.miccha.server.exception;
-
-import com.miccha.server.ErrorCode;
-
-public class DuplicateEmailException extends MicchaException {
-    @Override
-    public ErrorCode getErrorCode() {
-        return ErrorCode.DUPLICATE_EMAIL;
-    }
-}
