--- src/main/java/com/miccha/server/exception/MicchaException.java
@@ -1,7 +1,0 @@
-package com.miccha.server.exception;
-
-import com.miccha.server.ErrorCode;
-
-public abstract class MicchaException extends RuntimeException {
-    public abstract ErrorCode getErrorCode();
-}
