+++ src/main/java/com/miccha/server/exception/model/RequestMissingTokenException.java
@@ -1,0 +1,10 @@
+package com.miccha.server.exception.model;
+
+import com.miccha.server.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+public class RequestMissingTokenException extends MicchaException {
+    public RequestMissingTokenException() {
+        super(ErrorCode.REQUEST_MISSING_TOKEN, HttpStatus.BAD_REQUEST);
+    }
+}
